+++ drr/src/UI/src/app/app.routes.ts
@@ -3,7 +3,7 @@ import { AuthenticationGuard } from './core/guards/authentication.guard';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { DrifEoiViewComponent } from './drif/drif-eoi/drif-eoi-view/drif-eoi-view.component';
 import { EOIApplicationComponent } from './drif/drif-eoi/drif-eoi.component';
-import { DrifFpInstructionsComponent } from './drif/drif-fp-instructions/drif-fp-instructions.component';
+import { DrifFpInstructionsComponent } from './drif/drif-fp/drif-fp-instructions/drif-fp-instructions.component';
 import { DrifFpComponent } from './drif/drif-fp/drif-fp.component';
 import { SuccessPageComponent } from './success-page/success-page.component';
