+++ drr/src/UI/src/app/drif/drif-fp/drif-fp-step-5/drif-fp-step-5.component.html
@@ -118,7 +118,7 @@
           [id]="'otherFunding_description_' + i"
           [label]="t('otherDescription')"
           [rxFormControl]="
-            getArrayFormControl('otherDescription', 'otherFunding', i)
+            getArrayFormControl('otherFunding', 'otherDescription', i)
           "
           [maxlength]="100"
         ></drr-input>
