+++ drr/src/UI/src/app/drif/drif-fp/drif-fp-step-6/drif-fp-step-6.component.ts
@@ -4,9 +4,9 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatInputModule } from '@angular/material/input';
 import { TranslocoModule } from '@ngneat/transloco';
 import { IFormGroup } from '@rxweb/reactive-form-validators';
-import { DrrRadioButtonComponent } from '../../shared/controls/drr-radio-button/drr-radio-button.component';
-import { DrrTextareaComponent } from '../../shared/controls/drr-textarea/drr-textarea.component';
-import { ClimateAdaptationForm } from '../drif-fp/drif-fp-form';
+import { DrrRadioButtonComponent } from '../../../shared/controls/drr-radio-button/drr-radio-button.component';
+import { DrrTextareaComponent } from '../../../shared/controls/drr-textarea/drr-textarea.component';
+import { ClimateAdaptationForm } from '../drif-fp-form';
 
 @Component({
   selector: 'drif-fp-step-6',
