+++ drr/src/UI/src/app/shared/controls/drr-input/drr-input.component.ts
@@ -67,6 +67,9 @@ export class DrrInputComponent {
   @Input()
   set rxFormControl(rxFormControl: any) {
     this._formControl = rxFormControl as RxFormControl;
+    if (!rxFormControl) {
+      console.log(this.label);
+    }
   }
   get rxFormControl() {
     return this._formControl;
